chore: refactor ErrorBoundary component
